🚧 Isolate fast http in its own package - Move pong to `model` - Move api to `types`
