+++ internal/types/api.go
@@ -1,4 +1,4 @@
-package model
+package types
 
 import validation "github.com/go-ozzo/ozzo-validation"
